Unit tested transaction discovery via `--source parity_txqueue`
As above. Changed the Parity rpc endpoint to end with `/rpc` only to
make unit testing easier. Otherwise it would be difficult to mock the
`parity_pendingTransactions` endpoint leaving the original `testrpc`
communication in place.

Apparently, for real Parity node that change doesn't matter as it
responds to RPC requests regardless of the path.