Make nonce-reuse test work with Parity
This was a bit difficult to fix. I presume that
`data/parity/response.json` was intended to have the same data as the
Etherscan data, but it varied ever so slightly. The nonce on the
transactions from `OUR_ADDRESS` started at 8 in the Etherscan fixtures,
but starts at 9 in Parity's fixtures. In `get_pending_transactions()`
there is a filter for `tx.nonce > last_nonce`, and in this test,
`last_nonce` starts at 8.

This is part of removing support for Etherscan as a source and removing
the `--source` argument altogether.